% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chaos.r
\name{find_chaotic_borders}
\alias{find_chaotic_borders}
\title{Find the borders of chaotic motion from the results of test_chaos01_mw.}
\usage{
find_chaotic_borders(test01_res)
}
\arguments{
\item{test01_res}{The result of the 0-1 test for chaos calculated in mowing window.}
}
\value{
The list of optimized chaotic motion borders.
}
\description{
Find the borders of chaotic motion from the results of test_chaos01_mw.
}
