% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_regularity.r
\name{find_reg_borders}
\alias{find_reg_borders}
\title{Find the borders of regular motion from the results of test_chaos01_mw.}
\usage{
find_reg_borders(test01_res)
}
\arguments{
\item{test01_res}{The result of the 0-1 test for chaos calculated in mowing window.}
}
\value{
The list of optimized regular motion borders.
}
\description{
Find the borders of regular motion from the results of test_chaos01_mw.
}
