% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chaos.r
\name{optimize_chaos_run}
\alias{optimize_chaos_run}
\title{Optimization of chaotic motion borders based on bisection method.}
\usage{
optimize_chaos_run(find_thresh, test01_thresh = 0.05, chaos_borders_temp,
  data, skip_window, window_length, skip_test01)
}
\arguments{
\item{find_thresh}{Precision of found intervals.}

\item{test01_thresh}{The threshold to decide about motion.}

\item{chaos_borders_temp}{Borders of chaotic motion found by function find_chaotic_borders}

\item{data}{Analyzed data.}

\item{skip_window}{Length of the skip of the window moving in the data.}

\item{window_length}{Length of the window for in which the 0-1 test for chaos will be computed.}

\item{skip_test01}{Length of the skip to take data for calculation the 0-1 test for chaos in the window.}
}
\value{
The list of optimized chaotic motion borders.
}
\description{
Optimization of chaotic motion borders based on bisection method.
}
