% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_borders.R
\name{plot_borders}
\alias{plot_borders}
\title{Plot the chaotic and regular motions along with the data and results of the 0-1 test for chaos computed
using moving window.}
\usage{
plot_borders(data, window_length, skip_window, test01_res,
  chaos_borders_final, reg_borders_final)
}
\arguments{
\item{data}{Analyzed data.}

\item{window_length}{Length of the window for in which the 0-1 test for chaos will be computed.}

\item{skip_window}{Length of the skip of the window moving in the data.}

\item{test01_res}{The results of the 0-1 test for chaos in each computed window.}

\item{chaos_borders_final}{The found borders of chaotic motion.}

\item{reg_borders_final}{The found borders of regular motion.}
}
\description{
Plot the chaotic and regular motions along with the data and results of the 0-1 test for chaos computed
using moving window.
}
