% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chaos01_mw.R
\name{test_chaos01_mw}
\alias{test_chaos01_mw}
\title{Computes the 0-1 test for chaos in the window that is moving in the data.}
\usage{
test_chaos01_mw(data, window_length, skip_window, skip_test01,
  test01_thresh = 0.05)
}
\arguments{
\item{data}{Analyzed data.}

\item{window_length}{Length of the window for in which the 0-1 test for chaos will be computed}

\item{skip_window}{Length of the skip of the window moving in the data.}

\item{skip_test01}{Length of the skip to take data for calculation the 0-1 test for chaos in the window.}

\item{test01_thresh}{The threshold to decide about motion
#' \itemize{
  \item regular motion - results of the 0-1 test for chaos <= test01_thresh
  \item chaotic motion - results of the 0-1 test for chaos >= 1 - test01_thresh
  }}
}
\value{
The results of the 0-1 test for chaos in each computed window.
}
\description{
Computes the 0-1 test for chaos in the window that is moving in the data.
}
